--- src/main/java/rueppellii/backend2/tribes/building/BuildingRepository.java
@@ -1,6 +1,0 @@
-package rueppellii.backend2.tribes.building;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface BuildingRepository extends CrudRepository<Building, Long> {
-}
