+++ src/main/java/rueppellii/backend2/tribes/building/persistence/model/Building.java
@@ -1,9 +1,10 @@
-package rueppellii.backend2.tribes.building;
+package rueppellii.backend2.tribes.building.persistence.model;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import lombok.Getter;
 import lombok.Setter;
-import rueppellii.backend2.tribes.kingdom.Kingdom;
+import rueppellii.backend2.tribes.building.utility.BuildingType;
+import rueppellii.backend2.tribes.kingdom.persistence.model.Kingdom;
 
 import javax.persistence.*;
 import java.sql.Timestamp;
