+++ src/main/java/rueppellii/backend2/tribes/building/persistence/model/Farm.java
@@ -1,8 +1,8 @@
-package rueppellii.backend2.tribes.building;
+package rueppellii.backend2.tribes.building.persistence.model;
+
+import rueppellii.backend2.tribes.building.utility.BuildingType;
 
 import javax.persistence.Entity;
-import java.sql.Timestamp;
-import java.util.concurrent.TimeUnit;
 
 @Entity
 public class Farm extends Building {
