+++ src/main/java/rueppellii/backend2/tribes/building/persistence/model/Mine.java
@@ -1,4 +1,6 @@
-package rueppellii.backend2.tribes.building;
+package rueppellii.backend2.tribes.building.persistence.model;
+
+import rueppellii.backend2.tribes.building.utility.BuildingType;
 
 import javax.persistence.Entity;
 import java.sql.Timestamp;
