+++ src/main/java/rueppellii/backend2/tribes/building/utility/BuildingType.java
@@ -1,4 +1,6 @@
-package rueppellii.backend2.tribes.building;
+package rueppellii.backend2.tribes.building.utility;
+
+import rueppellii.backend2.tribes.building.persistence.model.*;
 
 public enum BuildingType {
