+++ src/main/java/rueppellii/backend2/tribes/building/web/BuildingController.java
@@ -1,10 +1,13 @@
-package rueppellii.backend2.tribes.building;
+package rueppellii.backend2.tribes.building.web;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import rueppellii.backend2.tribes.building.utility.BuildingDTO;
+import rueppellii.backend2.tribes.building.service.BuildingService;
+import rueppellii.backend2.tribes.building.persistence.model.Building;
 
 import java.security.Principal;
