+++ src/main/java/rueppellii/backend2/tribes/kingdom/persistence/repository/KingdomRepository.java
@@ -1,7 +1,8 @@
-package rueppellii.backend2.tribes.kingdom;
+package rueppellii.backend2.tribes.kingdom.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import rueppellii.backend2.tribes.kingdom.persistence.model.Kingdom;
 
 import java.util.Optional;
