+++ src/main/java/rueppellii/backend2/tribes/progression/persistence/ProgressionModel.java
@@ -3,7 +3,7 @@ package rueppellii.backend2.tribes.progression.persistence;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import lombok.Getter;
 import lombok.Setter;
-import rueppellii.backend2.tribes.kingdom.Kingdom;
+import rueppellii.backend2.tribes.kingdom.persistence.model.Kingdom;
 
 import javax.persistence.*;
