+++ src/main/java/rueppellii/backend2/tribes/user/persistence/model/ApplicationUser.java
@@ -3,7 +3,7 @@ package rueppellii.backend2.tribes.user.persistence.model;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import lombok.Getter;
 import lombok.Setter;
-import rueppellii.backend2.tribes.kingdom.Kingdom;
+import rueppellii.backend2.tribes.kingdom.persistence.model.Kingdom;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
